move types into separate file
